Move return_response function to utils.py
