+++ sdk/communication/azure-communication-chat/azure/communication/chat/_chat_client.py
@@ -15,14 +15,13 @@ from azure.core.tracing.decorator import distributed_trace
 from ._chat_thread_client import ChatThreadClient
 from ._common import CommunicationUserCredentialPolicy
 from ._shared.user_credential import CommunicationUserCredential
-from ._shared.response import return_response
 from ._generated import AzureCommunicationChatService
 from ._generated.models import CreateChatThreadRequest
 from ._models import (
     ChatThread,
     ChatThreadParticipant
 )
-from ._utils import _to_utc_datetime # pylint: disable=unused-import
+from ._utils import _to_utc_datetime, return_response # pylint: disable=unused-import
 from ._version import SDK_MONIKER
 
 if TYPE_CHECKING:
