--- sdk/communication/azure-communication-chat/azure/communication/chat/_shared/response.py
@@ -1,8 +1,0 @@
-# ------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for
-# license information.
-# -------------------------------------------------------------------------
-
-def return_response(response, deserialized, _):  # pylint: disable=unused-argument
-    return response, deserialized
